modify the blend factor to reduce the duplicate code and add blend factor for the motion-streak.
