+++ cocos2d/core/components/CCRenderComponent.js
@@ -26,7 +26,78 @@
 const Component = require('./CCComponent');
 const renderEngine = require('../renderer/render-engine');
 const RenderFlow = require('../renderer/render-flow');
+const BlendFactor = require('../platform/CCMacro').BlendFactor;
 const RenderData = renderEngine.RenderData;
+const gfx = renderEngine.gfx;
+
+let BlendFunction = cc.Class({
+    name: 'BlendFunction',
+
+    properties: {
+        _srcBlendFactor: BlendFactor.SRC_ALPHA,
+        _dstBlendFactor: BlendFactor.ONE_MINUS_SRC_ALPHA,
+
+          /**
+         * !#en specify the source Blend Factor, this will generate a custom material object, please pay attention to the memory cost.
+         * !#zh 指定原图的混合模式，这会克隆一个新的材质对象，注意这带来的开销
+         * @property srcBlendFactor
+         * @type {macro.BlendFactor}
+         * @example
+         * sprite.srcBlendFactor = cc.macro.BlendFactor.ONE;
+         */
+        srcBlendFactor: {
+            get: function() {
+                return this._srcBlendFactor;
+            },
+            set: function(value) {
+                if (this._srcBlendFactor === value) return;
+                this._srcBlendFactor = value;
+                this._updateBlendFunc(true);
+            },
+            animatable: false,
+            type:BlendFactor,
+            tooltip: CC_DEV && 'i18n:COMPONENT.sprite.src_blend_factor'
+        },
+
+        /**
+         * !#en specify the destination Blend Factor.
+         * !#zh 指定目标的混合模式
+         * @property dstBlendFactor
+         * @type {macro.BlendFactor}
+         * @example
+         * sprite.dstBlendFactor = cc.macro.BlendFactor.ONE;
+         */
+        dstBlendFactor: {
+            get: function() {
+                return this._dstBlendFactor;
+            },
+            set: function(value) {
+                if (this._dstBlendFactor === value) return;
+                this._dstBlendFactor = value;
+                this._updateBlendFunc(true);
+            },
+            animatable: false,
+            type: BlendFactor,
+            tooltip: CC_DEV && 'i18n:COMPONENT.sprite.dst_blend_factor'
+        },
+    },
+
+    _updateBlendFunc: function (updateHash) {
+        let material = this.getMaterial();
+        if (material) {
+            var pass = material._mainTech.passes[0];
+            pass.setBlend(
+                gfx.BLEND_FUNC_ADD,
+                this._srcBlendFactor, this._dstBlendFactor,
+                gfx.BLEND_FUNC_ADD,
+                this._srcBlendFactor, this._dstBlendFactor
+            );
+            if (updateHash) {
+                material.updateHash();
+            }
+        }
+    },
+})
 
 /**
  * !#en
@@ -156,5 +227,6 @@ let RenderComponent = cc.Class({
 });
 RenderComponent._assembler = null;
 RenderComponent._postAssembler = null;
+RenderComponent.BlendFunction = BlendFunction;
 
 cc.RenderComponent = module.exports = RenderComponent;
